+++ src/com/dci/intellij/dbn/editor/data/DatasetEditor.java
@@ -287,7 +287,7 @@ public class DatasetEditor extends UserDataHolderBase implements FileEditor, Fil
 
     public void loadData(final DatasetLoadInstructions instructions) {
         if (status.isNot(LOADING)) {
-            ConnectionAction.invoke("loading table data", this, (Integer) null, action -> {
+            ConnectionAction.invoke("loading table data", this, null, action -> {
                 setLoading(true);
                 EventUtil.notify(project, DatasetLoadListener.TOPIC).datasetLoading(databaseFile);
                 SimpleBackgroundTask.invoke(() -> {
